==> 01_IMDB_webscrapper_sequels_franchise_Part_I.ipynb <==
## Scrape IMDB Movies with Sequels/Franchise (Part I)

As part of revamping Project 2 and potentially for Springboard, I am re-scrapping the sequels movie page on IMDB. This will hopefully be better content going forward.

This notebook only scrapes the movie list on 

https://www.imdb.com/list/ls003495084/?st_dt=&mode=detail&page=1&sort=list_order,asc

## Load Libraries

from bs4 import BeautifulSoup as soup  # HTML data structure
from urllib.request import urlopen as uReq  # Web client

import pandas as pd
import numpy as np

from tqdm import tqdm

## Start Scrapping

Main page url is: https://www.imdb.com/list/ls003495084/?st_dt=&mode=detail&page=1&sort=list_order,asc

The idea here is as follows:

1. Scrape the urls of all movies on the sequels page
2. Use those urls to go to their respective pages and parse necessary data

The challenge here will be to scroll through pages and grab all data. Not all data will be available. Example, James Bond and I found out that Transformers: The Last Knight is not on that page at all.

def get_movie_sequels_urls_n_more(num_pages=12):
    
    # Initialize lists
    movie_url        = []
    movie_title      = []
    movie_year       = []
    movie_imdb_score = []
    movie_meta_score = []
    movie_rating     = []
    movie_genre      = []
    movie_runtime    = []
    
    # Function to get main page urls and extra information
    for ipage in range(1,num_pages+1):
        
        print('Start: ', ipage)
        
        # This is the list of movie sequels page
        main_page_url =  'https://www.imdb.com/list/ls003495084/?st_dt=&mode=detail&page='\
                      +   str(ipage) + '&sort=list_order,asc'
        
        # Opens the connection and downloads html page from url
        uClient = uReq(main_page_url)
        
        # Parses html into a soup data structure to traverse html
        # as if it were a json data type.
        page_soup = soup(uClient.read(), "html.parser")
        uClient.close()
        
        # Now get all movie containers (typically should total 100)
        containers = page_soup.find_all('div', class_ = 'lister-item-content')
        
        # Loop through all the containers
        
        for container in containers:
            # Grab url link
            try:
                m_url = 'http://www.imdb.com' + container.a['href']
                movie_url.append(m_url)
            except Exception as e:
                movie_url.append(np.nan)
            
            # Grab movie title
            try:
                movie_title.append(container.a.text)
            except Exception as e:
                movie_title.append(np.nan)
                
            # Grab the movie year which will also include whether or not this was a video release
            try:
                m_year = container.h3.find('span', class_ = 'lister-item-year text-muted unbold').text.strip()
                m_year = m_year.replace('(','')
                m_year = m_year.replace(')','')
                m_year = m_year.replace('I','')
                
                if m_year != '':
                    movie_year.append(m_year.strip())
                else:
                    movie_year.append(np.nan)
            except Exception as e:
                movie_year.append(np.nan)
            
            # Grab IMDB score
            try:
                m_imdb_score = container.find('span', class_ = 'ipl-rating-star__rating').text
                movie_imdb_score.append(m_imdb_score)
            except Exception as e:
                movie_imdb_score.append(np.nan)
                
            # Grab Metacritic score
            try:
                m_meta_score = container.find('div', class_ = 'inline-block ratings-metascore').span.text
                movie_meta_score.append(m_meta_score.strip())
            except Exception as e:
                movie_meta_score.append(np.nan)
                
            # Now extract rating, runtime and genre
            top_banner = container.find('p', class_ = 'text-muted text-small')
            
            try:
                m_rating = top_banner.find('span', class_='certificate').text.strip()
                movie_rating.append(m_rating)
            except Exception as e:
                movie_rating.append(np.nan)
    
            try:
                runtime = top_banner.find('span', class_='runtime').text.strip()
                runtime = int(runtime.split(' ')[0])
                movie_runtime.append(runtime)
            except Exception as e:
                movie_runtime.append(np.nan)
    
            try:
                m_genre = top_banner.find('span', class_='genre').text.strip()
                movie_genre.append(''.join(m_genre.split(',')))
            except Exception as e:
                movie_genre.append(np.nan)
                
        print('End: ', ipage)
                
    # Create a dictionary and dump it to a dataframe
    movie_dict = {'Title'            : movie_title,
                  'url'              : movie_url, 
                  'Year-Rel-Type'    : movie_year,
                  'IMDB Score'       : movie_imdb_score,
                  'Metacritic Score' : movie_meta_score,
                  'Rating'           : movie_rating,
                  'Genre'            : movie_genre,
                  'Runtime'          : movie_runtime}
        
    df_movies = pd.DataFrame(movie_dict)
                
    return(df_movies)
                
    # end for loop ipage



num_page_test = 12
dfm = get_movie_sequels_urls_n_more(num_page_test)

dfm.iloc[499,:]

# Dump to a csv file
dfm.to_csv('./data/movies_with_sequels_imdb_first_pass.csv',index=False)

dfm1 = pd.read_csv('./data/movies_with_sequels_imdb_first_pass.csv')

dfm1.head()

# Need a way to break-up the categorical data in Genre
genre = dfm['Genre']

#print(list(genre))

master = []

for item in genre:
    
    list_genre = item.split(' ')
    
    master.extend(list_genre)
    
unique_genres = sorted(list(set(master)))

print(len(unique_genres))
print(unique_genres)

dfm_test = dfm.iloc[:10,:]

#dfm_test_copy = dfm_test.copy()

#dfm_test_copy['Action']    = 0
#dfm_test_copy['Adventure'] = 0
#dfm_test_copy['Sci-Fi']    = 0

dfm_test.head(10)

dfm_new = dfm_test.drop(['Title','url'],axis=1)

dfm_new_copy = dfm_new.copy()

for gg in unique_genres:
    dfm_new_copy[gg] = 0
    
dfm_new_copy.head(10)

## Scrape James Bond URLs

# Bond movies are on a separate list page on IMDB

def get_movie_sequels_urls_n_more_bond(num_pages=1):
    
    # Initialize lists
    movie_url        = []
    movie_title      = []
    movie_year       = []
    movie_imdb_score = []
    movie_meta_score = []
    movie_rating     = []
    movie_genre      = []
    movie_runtime    = []
    
    # Function to get main page urls and extra information
    for ipage in range(1,num_pages+1):
        
        print('Start: ', ipage)
        
        # # This is the James Bond Page
        main_page_url =  'https://www.imdb.com/list/ls006405458/'
        
        # Opens the connection and downloads html page from url
        uClient = uReq(main_page_url)
        
        # Parses html into a soup data structure to traverse html
        # as if it were a json data type.
        page_soup = soup(uClient.read(), "html.parser")
        uClient.close()
        
        # Now get all movie containers (typically should total 100)
        containers = page_soup.find_all('div', class_ = 'lister-item-content')
        
        # Loop through all the containers
        
        for container in containers:
            # Grab url link
            try:
                m_url = 'http://www.imdb.com' + container.a['href']
                movie_url.append(m_url)
            except Exception as e:
                movie_url.append(np.nan)
            
            # Grab movie title
            try:
                movie_title.append(container.a.text)
            except Exception as e:
                movie_title.append(np.nan)
                
            # Grab the movie year which will also include whether or not this was a video release
            try:
                m_year = container.h3.find('span', class_ = 'lister-item-year text-muted unbold').text.strip()
                m_year = m_year.replace('(','')
                m_year = m_year.replace(')','')
                m_year = m_year.replace('I','')
                
                if m_year != '':
                    movie_year.append(m_year.strip())
                else:
                    movie_year.append(np.nan)
            except Exception as e:
                movie_year.append(np.nan)
            
            # Grab IMDB score
            try:
                m_imdb_score = container.find('span', class_ = 'ipl-rating-star__rating').text
                movie_imdb_score.append(m_imdb_score)
            except Exception as e:
                movie_imdb_score.append(np.nan)
                
            # Grab Metacritic score
            try:
                m_meta_score = container.find('div', class_ = 'inline-block ratings-metascore').span.text
                movie_meta_score.append(m_meta_score.strip())
            except Exception as e:
                movie_meta_score.append(np.nan)
                
            # Now extract rating, runtime and genre
            top_banner = container.find('p', class_ = 'text-muted text-small')
            
            try:
                m_rating = top_banner.find('span', class_='certificate').text.strip()
                movie_rating.append(m_rating)
            except Exception as e:
                movie_rating.append(np.nan)
    
            try:
                runtime = top_banner.find('span', class_='runtime').text.strip()
                runtime = int(runtime.split(' ')[0])
                movie_runtime.append(runtime)
            except Exception as e:
                movie_runtime.append(np.nan)
    
            try:
                m_genre = top_banner.find('span', class_='genre').text.strip()
                movie_genre.append(''.join(m_genre.split(',')))
            except Exception as e:
                movie_genre.append(np.nan)
                
        print('End: ', ipage)
                
    # Create a dictionary and dump it to a dataframe
    movie_dict = {'Title'            : movie_title,
                  'url'              : movie_url, 
                  'Year-Rel-Type'    : movie_year,
                  'IMDB Score'       : movie_imdb_score,
                  'Metacritic Score' : movie_meta_score,
                  'Rating'           : movie_rating,
                  'Genre'            : movie_genre,
                  'Runtime'          : movie_runtime}
        
    df_movies = pd.DataFrame(movie_dict)
                
    return(df_movies)


df_mo_bond = get_movie_sequels_urls_n_more_bond(1)

df_mo_bond.to_csv('./data/movies_with_sequels_imdb_first_pass_james_bond.csv',index=False)

## Scrape Madea URLs

def get_movie_sequels_urls_n_more_madea(num_pages=1):
    
    # Initialize lists
    movie_url        = []
    movie_title      = []
    movie_year       = []
    movie_imdb_score = []
    movie_meta_score = []
    movie_rating     = []
    movie_genre      = []
    movie_runtime    = []
    
    # Function to get main page urls and extra information
    for ipage in range(1,num_pages+1):
        
        print('Start: ', ipage)
        
        # This is the Madea Page
        main_page_url =  'https://www.imdb.com/search/keyword/?keywords=madea-series&ref_=fn_al_kw_1&sort=release_date,asc&mode=detail&page=1'
        
        # Opens the connection and downloads html page from url
        uClient = uReq(main_page_url)
        
        # Parses html into a soup data structure to traverse html
        # as if it were a json data type.
        page_soup = soup(uClient.read(), "html.parser")
        uClient.close()
        
        # Now get all movie containers (typically should total 100)
        containers = page_soup.find_all('div', class_ = 'lister-item-content')
        
        # Loop through all the containers
        
        for container in containers:
            # Grab url link
            try:
                m_url = 'http://www.imdb.com' + container.a['href']
                movie_url.append(m_url)
            except Exception as e:
                movie_url.append(np.nan)
            
            # Grab movie title
            try:
                movie_title.append(container.a.text)
            except Exception as e:
                movie_title.append(np.nan)
                
            # Grab the movie year which will also include whether or not this was a video release
            try:
                m_year = container.h3.find('span', class_ = 'lister-item-year text-muted unbold').text.strip()
                m_year = m_year.replace('(','')
                m_year = m_year.replace(')','')
                m_year = m_year.replace('I','')
                
                if m_year != '':
                    movie_year.append(m_year.strip())
                else:
                    movie_year.append(np.nan)
            except Exception as e:
                movie_year.append(np.nan)
                
            try:
                # Slightly different for Madea
                m_imdb_score = container.find('div', class_ = 'inline-block ratings-imdb-rating').strong.text
                movie_imdb_score.append(m_imdb_score)
            except Exception as e:
                movie_imdb_score.append(np.nan)
                
            # Grab Metacritic score
            try:
                m_meta_score = container.find('div', class_ = 'inline-block ratings-metascore').span.text
                movie_meta_score.append(m_meta_score.strip())
            except Exception as e:
                movie_meta_score.append(np.nan)
                
            # Now extract rating, runtime and genre
            top_banner = container.find('p', class_ = 'text-muted text-small')
            
            try:
                m_rating = top_banner.find('span', class_='certificate').text.strip()
                movie_rating.append(m_rating)
            except Exception as e:
                movie_rating.append(np.nan)
    
            try:
                runtime = top_banner.find('span', class_='runtime').text.strip()
                runtime = int(runtime.split(' ')[0])
                movie_runtime.append(runtime)
            except Exception as e:
                movie_runtime.append(np.nan)
    
            try:
                m_genre = top_banner.find('span', class_='genre').text.strip()
                movie_genre.append(''.join(m_genre.split(',')))
            except Exception as e:
                movie_genre.append(np.nan)
                
        print('End: ', ipage)
                
    # Create a dictionary and dump it to a dataframe
    movie_dict = {'Title'            : movie_title,
                  'url'              : movie_url, 
                  'Year-Rel-Type'    : movie_year,
                  'IMDB Score'       : movie_imdb_score,
                  'Metacritic Score' : movie_meta_score,
                  'Rating'           : movie_rating,
                  'Genre'            : movie_genre,
                  'Runtime'          : movie_runtime}
        
    df_movies = pd.DataFrame(movie_dict)
                
    return(df_movies)

df_mo_madea = get_movie_sequels_urls_n_more_madea(1)

df_mo_madea.to_csv('./data/movies_with_sequels_imdb_first_pass_madea.csv',index=False)

## Scrape Some Bollywood Movies

Bollywood movies contain a wealth of information and it is perhaps the second biggest movie market size in the world behind Hollywood. The list on IMDB is old and is way short of comprehensive.

https://www.imdb.com/list/ls000010248/

def get_movie_sequels_urls_n_more_bollywood(num_pages=1):
    
    # Initialize lists
    movie_url        = []
    movie_title      = []
    movie_year       = []
    movie_imdb_score = []
    movie_meta_score = []
    movie_rating     = []
    movie_genre      = []
    movie_runtime    = []
    
    # Function to get main page urls and extra information
    for ipage in range(1,num_pages+1):
        
        print('Start: ', ipage)
        
        # This is Bollywood page on IMDB
        # Severely incomplete
        main_page_url =  'https://www.imdb.com/list/ls000010248/'
        
        # Opens the connection and downloads html page from url
        uClient = uReq(main_page_url)
        
        # Parses html into a soup data structure to traverse html
        # as if it were a json data type.
        page_soup = soup(uClient.read(), "html.parser")
        uClient.close()
        
        # Now get all movie containers (typically should total 100)
        containers = page_soup.find_all('div', class_ = 'lister-item-content')
        
        # Loop through all the containers
        
        for container in containers:
            # Grab url link
            try:
                m_url = 'http://www.imdb.com' + container.a['href']
                movie_url.append(m_url)
            except Exception as e:
                movie_url.append(np.nan)
            
            # Grab movie title
            try:
                movie_title.append(container.a.text)
            except Exception as e:
                movie_title.append(np.nan)
                
            # Grab the movie year which will also include whether or not this was a video release
            try:
                m_year = container.h3.find('span', class_ = 'lister-item-year text-muted unbold').text.strip()
                m_year = m_year.replace('(','')
                m_year = m_year.replace(')','')
                m_year = m_year.replace('I','')
                
                if m_year != '':
                    movie_year.append(m_year.strip())
                else:
                    movie_year.append(np.nan)
            except Exception as e:
                movie_year.append(np.nan)
            
            # Get IMDB score
            try:
                # Slightly different for Madea
                m_imdb_score = container.find('span', class_ = 'ipl-rating-star__rating').text
                movie_imdb_score.append(m_imdb_score)
            except Exception as e:
                movie_imdb_score.append(np.nan)
                
            # Grab Metacritic score
            try:
                m_meta_score = container.find('div', class_ = 'inline-block ratings-metascore').span.text
                movie_meta_score.append(m_meta_score.strip())
            except Exception as e:
                movie_meta_score.append(np.nan)
                
            # Now extract rating, runtime and genre
            top_banner = container.find('p', class_ = 'text-muted text-small')
            
            try:
                m_rating = top_banner.find('span', class_='certificate').text.strip()
                movie_rating.append(m_rating)
            except Exception as e:
                movie_rating.append(np.nan)
    
            try:
                runtime = top_banner.find('span', class_='runtime').text.strip()
                runtime = int(runtime.split(' ')[0])
                movie_runtime.append(runtime)
            except Exception as e:
                movie_runtime.append(np.nan)
    
            try:
                m_genre = top_banner.find('span', class_='genre').text.strip()
                movie_genre.append(''.join(m_genre.split(',')))
            except Exception as e:
                movie_genre.append(np.nan)
                
        print('End: ', ipage)
                
    # Create a dictionary and dump it to a dataframe
    movie_dict = {'Title'            : movie_title,
                  'url'              : movie_url, 
                  'Year-Rel-Type'    : movie_year,
                  'IMDB Score'       : movie_imdb_score,
                  'Metacritic Score' : movie_meta_score,
                  'Rating'           : movie_rating,
                  'Genre'            : movie_genre,
                  'Runtime'          : movie_runtime}
        
    df_movies = pd.DataFrame(movie_dict)
                
    return(df_movies)

df_mo_bolly = get_movie_sequels_urls_n_more_bollywood(1)

df_mo_bolly.head(5)

df_mo_bolly.to_csv('./data/movies_with_sequels_imdb_first_pass_bolly.csv',index=False)
