==> 03_Box_Office_Movie_Franchise_Predictor_Cleaning.ipynb <==
# Box Office Movie Franchise Predictor - Cleaning

Before embarking on this, a lot of time was spent searching for these movies. Because the number of movies that belong in a sequel fashion is limited, the scope of the study also includes movies that belong to a franchise and universe. A franchise example is James Bond and one that belongs to a universe is the Marvel Cinematic Universe (MCU). 

The time sinks in gathering and cleaning the data were:

1. Release Dates - Some formats were incomplete and had to be researched in either IMDB and cross-referenced with Wikipedia


2. Budgets & Currency - Missing budgets and foreign currency had to be converted accordingly to USD based on the rough exchange rate on the day of release


3. Correcting dates that were 1968 and earlier. Python parses the dates in the csv file a little strange and gave dates that are in 2068 for example. This is corrected.


## Import Libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import datetime

## Load Data File(s)

data_path = './data/'
file_name = 'movies_with_sequels_final.csv'

df = pd.read_csv(data_path + file_name)

## Data Wrangling & Cleaning

df.head(6)

### A Note about the Dataset

One very important note on how the csv file is arranged. The entire dataset as one can see from the above, i.e. an original movie or first movie and its corresponding sequel is arranged such that 

1. Movie 1
2. Sequel of Movie 1
3. Another film
4. Sequel of Another Film
5. ...
6. ...

The first 6 rows of the Pandas data frame (above), the movie is arranged as discussed. However, there was a Spider-Man 3 but it is not included here, i.e. index 2. The structure here has to be in even numbers and trilogies are not considered here for the time being. If the dataset is to be absent of leaky variables, the structure of observations must be in even numbers. That way, after cleaning, we can collect every other row as a target without the fear of a leaky variable.

df.info()

From the info table above, there are no missing values. This is incorrect. There are missing values but are stated as 'None' in their respective columns and that is why an entire column gets converted to an object. This was done initially in the webscrapping portion of the work. It has been corrected to be replaced with np.nan. Nonetheless, the change will be made here.

### Replace 'None' with NaN

def replace_none_w_nan(data_frame):
    
    column_names = list(data_frame.columns)
    
    for col in column_names:
        data_frame[col].replace('None',np.nan, inplace=True)
        
    return data_frame


df = replace_none_w_nan(df)

df.info()

This is much better. There are two missing budget entries which can be fixed. Too many missing values for

- Metacritic Score
- Opening Weekend
- Gross USA

First let's drop the Opening Weekend and Gross USA columns as there is too much variability in those. The Opening Weekend metric was not really tallied only until the late 90's. Let's see what we can do with the Metacritic and two missing Budgets. The missing budgets are from the movies Homeward Bound 1 and 2. But first, we need to clean them. 

### Clean-Up Budget

def money_to_float(money_str):
    
    money_str = str(money_str)
    money_str = money_str.strip()
    
    return float(money_str.replace('$','').replace(',',''))


# Clean money and convert to float
df['Budget'] = df['Budget'].apply(money_to_float)
df['Opening Weekend'] = df['Opening Weekend'].apply(money_to_float)
df['Gross USA'] = df['Gross USA'].apply(money_to_float)
df['Gross World'] = df['Gross World'].apply(money_to_float)

df.info()

df.loc[346]

### Clean & Fix Release Dates

There is an apparent problem with Python when handling dates from 1968 or 1969. Apparently, if the year is 3-Nov-68, it converts it to 2068-11-03. Hence, a fix is needed for this.

# Convert Release Date to proper date
# This is an example showing the problem
df['Release Date'] = pd.to_datetime(df['Release Date']) #,format='%d-%b-%y')

df.loc[346]

def fix_dates(rel_date):
    
    # Dates should not go this far
    cut_off = 2030
    
    if rel_date.year > cut_off:    
        year = rel_date.year - 100
    else:
        year = rel_date.year
    
    return datetime.date(year,rel_date.month,rel_date.day)

df['Release Date'] = df['Release Date'].apply(fix_dates)

df.head(6)

### Write to a New File

# Dump to new csv
file_name_cleaned = 'movies_with_sequels_final_cleaned.csv'

df.to_csv(data_path + file_name_cleaned, index=False)


## Splitting the Dataset to 'Original' and 'Sequel'

# Dump original and sequel
df_original = df[df.index % 2 == 0] # Excludes every 3rd row starting from 0
df_sequel   = df[df.index % 2 != 0] # Selects every 3rd raw starting from 0

df_original.reset_index(drop=True,inplace=True)

df_sequel.reset_index(drop=True,inplace=True)

df_original.head()

df_sequel.head()

# Dump to new files
file_name_cleaned_org = 'movies_with_originals_only_final_cleaned.csv'
file_name_cleaned_seq = 'movies_with_sequels_only_final_cleaned.csv'

df_original.to_csv(data_path + file_name_cleaned_org, index=False)
df_sequel.to_csv(data_path + file_name_cleaned_seq, index=False)

df_original.info()

df_sequel.info()

### Check for Negative Difference in Release Dates

A sanity check to see if there are negative days between original and release. This was done since I noticed there were delta days that either negative or really large. Function below will be used again in the next notebook.

# Let's check and see if there are any negative deltas between original and sequel
def diff_rel_dates(data_org,data_seq,delta_type):
    
    df_diff_days = (data_seq['Release Date'] - data_org['Release Date'])/np.timedelta64(1, delta_type)
    
    neg_diff_days = df_diff_days[df_diff_days < 0]
    
    if neg_diff_days is not None:
        print('No negative difference')
    else:
        print('There are negative time(s) between sequel and original. Please check!')

    return df_diff_days

diff_type = 'D'  # Let's use days as a difference

diff_days = diff_rel_dates(df_original,df_sequel,diff_type)


Great! There are no negative delta days between original and sequel.
